делает раздел 5
