+++ index.html
@@ -234,6 +234,6 @@
       </section>
     </template>
 
-  <script src="functions.js"></script>
+  <script src="main.js" type="module"></script>
 </body>
 </html>
