+++ js/data.js
@@ -1,0 +1,21 @@
+//то что известно до начала выполнения задания
+
+const NAMES = [
+  'Артем', 'Светлана', 'Игорь', 'Мария', 'Ксения', 'Петр', 'Иван', 'Владимир','Алексей', 'Ассоль', 'Лилия', 'Максим', 'Константин', 'Евгения', 'Михаил', 'Станислав', 'Елена', 'Александр', 'Марина', 'Вероника', 'Бен', 'Стен', 'Глен', 'Ален', 'Мирослава',
+];
+
+const MESSAGES = [
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  'Всё отлично!',
+];
+
+const DESCRIPTIONS = [
+  'Иконка с буквой К на коричневом фоне',
+  'Иконка колокола на зеленом фоне',
+  'Иконка с мороженным на фиолетовом фоне',
+  'Иконка с бабочкой на розовом фоне',
+  'Иконка снежинки на желтом фоне',
+  'Иконка грузовика на голубом фоне',
+];
+
+export {NAMES, MESSAGES, DESCRIPTIONS};
