+++ js/functions.js
@@ -17,5 +17,3 @@ const isPalindrome = (string) => {
 };
 
 isPalindrome ();
-
-
